+++ public/utils/CRC32.cpp
@@ -19,7 +19,7 @@ void CRC32_Update( uint32 &crcvalue, const char data )
 	crcvalue = crc32_detail::crctable[ ( crcvalue ^ data ) & 0xff ] ^ ( crcvalue >> 8 );
 }
 
-void CRC32_UpdateChecksum( uint32 &crcvalue, const void *data, int length )
+void CRC32_UpdateChecksum( uint32 &crcvalue, const void *data, size_t length )
 {
 	uint32 crc;
 	const unsigned char *buf = (const unsigned char *) data;
@@ -39,7 +39,7 @@ void CRC32_FinishChecksum( uint32 &crcvalue )
 	crcvalue ^= CRC32_XOR_VALUE;
 }
 
-uint32 CRC32_BlockChecksum( const void *data, int length )
+uint32 CRC32_BlockChecksum( const void *data, size_t length )
 {
 	uint32 crc;
